refactor(memory): extract pure ops into memory-ops module
