+++ scripts/memory-bench-scenarios.ts
@@ -2,8 +2,9 @@ import { existsSync } from "node:fs";
 import { mkdir, readFile, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { z } from "zod";
-import { DISTILLER_PROMPT, splitScopedObservation } from "../src/memory-distiller";
+import { DISTILLER_PROMPT } from "../src/memory-distiller";
 import { cosineSimilarity, embedText } from "../src/memory-embedding";
+import { splitScopedObservation } from "../src/memory-ops";
 
 export type MemoryBenchDatasetId = "longmemeval" | "locomo" | "locomo-observations" | "locomo-distilled";
