+++ src/memory-distiller.test.ts
@@ -1,12 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import type { MemoryKind, MemoryRecord, MemoryStore } from "./memory-contract";
 import { createMemoryPolicy } from "./memory-contract";
-import {
-  clampToTokenEstimate,
-  createMemoryDistiller,
-  DISTILLER_PROMPT,
-  splitScopedObservation,
-} from "./memory-distiller";
+import { createMemoryDistiller, DISTILLER_PROMPT } from "./memory-distiller";
 
 const testPolicy = createMemoryPolicy({ messageThreshold: 1, maxOutputTokens: 200 });
 
@@ -50,66 +45,6 @@ function createMockStore(records: MemoryRecord[] = []): MemoryStore & { written:
     close() {},
   };
 }
-
-describe("clampToTokenEstimate", () => {
-  test("does not produce lone surrogates when clamping emoji text", () => {
-    // Mix single-byte and surrogate pair chars to force odd-boundary slicing.
-    // "a🎉" = 3 UTF-16 code units. Repeating gives lengths not divisible by 2.
-    const mixed = "a🎉".repeat(200);
-    const result = clampToTokenEstimate(mixed, 5);
-    expect(result.length).toBeGreaterThan(0);
-    expect(result).not.toMatch(/[\uD800-\uDBFF](?![\uDC00-\uDFFF])/);
-    expect(result).not.toMatch(/(?<![\uD800-\uDBFF])[\uDC00-\uDFFF]/);
-  });
-});
-
-describe("splitScopedObservation", () => {
-  const split = splitScopedObservation;
-
-  test("parses scoped observations into facts", () => {
-    const result = split("@observe project\nuses bun\n@observe user\nprefers terse output");
-    expect(result.projectCount).toBe(1);
-    expect(result.userCount).toBe(1);
-    expect(result.facts).toHaveLength(2);
-    expect(result.facts[0]).toMatchObject({ scope: "project", content: "uses bun" });
-  });
-
-  test("drops untagged lines without a preceding @observe", () => {
-    const result = split("orphan line\n@observe session\ntagged line");
-    expect(result.droppedUntaggedCount).toBe(1);
-    expect(result.sessionCount).toBe(1);
-  });
-
-  test("drops malformed @observe directives", () => {
-    const result = split("@observe badscope\nfollowing line");
-    expect(result.droppedMalformedCount).toBe(1);
-    expect(result.droppedUntaggedCount).toBe(1);
-    expect(result.facts).toHaveLength(0);
-  });
-
-  test("handles @observe at end of input with no content", () => {
-    const result = split("@observe project");
-    expect(result.facts).toHaveLength(0);
-    expect(result.projectCount).toBe(0);
-  });
-
-  test("handles @topic directive", () => {
-    const result = split("@observe project\n@topic testing\nuses bun test");
-    expect(result.facts[0]).toMatchObject({ scope: "project", content: "uses bun test", topic: "testing" });
-  });
-
-  test("@topic without preceding @observe — content is dropped as untagged", () => {
-    const result = split("@topic orphan\nsome content");
-    expect(result.droppedUntaggedCount).toBe(1);
-    expect(result.facts).toHaveLength(0);
-  });
-
-  test("multiple @observe in sequence — only last one applies", () => {
-    const result = split("@observe project\n@observe user\nactual fact");
-    expect(result.userCount).toBe(1);
-    expect(result.projectCount).toBe(0);
-  });
-});
 
 describe("memoryDistiller", () => {
   describe("commit", () => {
